+++ src/app/collections/correctiveAction/correctiveAction.controller.js
@@ -2,10 +2,10 @@
     'use strict';
 
     angular.module('chpl.collections')
-        .controller('NonconformitiesController', NonconformitiesController);
+        .controller('CorrectiveActionController', CorrectiveActionController);
 
     /** @ngInject */
-    function NonconformitiesController () {
+    function CorrectiveActionController () {
         var vm = this;
 
         activate();
@@ -23,12 +23,22 @@
                 { predicate: 'openNonconformityCount', display: '# Open NCs', sortType: 'multi', descendingFirst: true, sortDefault: ['-openNonconformityCount', '-edition', 'developer'] },
                 { predicate: 'closedNonconformityCount', display: '# Closed NCs', sortType: 'multi', descendingFirst: true },
             ];
-            vm.filters = ['acb', 'edition'];
+            vm.filters = ['acb', 'certificationStatus', 'edition', 'nonconformities'];
             vm.refineModel = {
                 acb: [
                     { value: 'Drummond Group', selected: true },
                     { value: 'ICSA Labs', selected: true },
                     { value: 'InfoGard', selected: true },
+                ],
+                certificationStatus: [
+                    { value: 'Active', selected: true },
+                    { value: 'Suspended by ONC', selected: true },
+                    { value: 'Suspended by ONC-ACB', selected: true },
+                    { value: 'Retired', selected: false },
+                    { value: 'Withdrawn by Developer', selected: false },
+                    { value: 'Withdrawn by Developer Under Surveillance/Review', selected: false },
+                    { value: 'Withdrawn by ONC-ACB', selected: false },
+                    { value: 'Terminated by ONC', selected: false },
                 ],
                 edition: [
                     { value: '2014', selected: true },